pyproject.toml
[tool.poetry]
name = "jeru"
version = "0.1.0"
description = "Price comparator of Supermarket Products"
authors = ["Victor Pereira"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.2"
Flask-CLI = "^0.4.0"
Flask-SQLAlchemy = "^2.5.1"
python-decouple = "^3.5"
SQLAlchemy-serializer = "^1.4.1"
bcrypt = "^3.2.0"
Flask-HTTPAuth = "^4.5.0"
Scrapy = "^2.5.1"
celery = "^5.2.3"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
